% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESAAdmittedAggregated.R
\name{ESAAdmittedAggregated}
\alias{ESAAdmittedAggregated}
\title{Title: ESAAdmittedAggregated
Author: Edmund Haacke
Date created: 22.09.2021
Date modified: 22.09.2021
Changelog:
- 22.09.21: file created.
Description:
Container for the admitted patient care episodes.
Transfers calculation
4 definition of transfers used here
A) burden on site where transfer arrived at
A.1) where an episode moves only specialty code
A.2) where an episode moves specialty code and site
A.3) where an episode moves only site (but same specialty)
B) burden on site where transfer originated from
B.1) where an episode arrives to a new site, irrrespective of specialty}
\description{
Title: ESAAdmittedAggregated
Author: Edmund Haacke
Date created: 22.09.2021
Date modified: 22.09.2021
Changelog:
- 22.09.21: file created.
Description:
Container for the admitted patient care episodes.
Transfers calculation
4 definition of transfers used here
A) burden on site where transfer arrived at
A.1) where an episode moves only specialty code
A.2) where an episode moves specialty code and site
A.3) where an episode moves only site (but same specialty)
B) burden on site where transfer originated from
B.1) where an episode arrives to a new site, irrrespective of specialty

Title: ESAAdmittedAggregated
Author: Edmund Haacke
Date created: 22.09.2021
Date modified: 22.09.2021
Changelog:
- 22.09.21: file created.
Description:
Container for the admitted patient care episodes.
Transfers calculation
4 definition of transfers used here
A) burden on site where transfer arrived at
A.1) where an episode moves only specialty code
A.2) where an episode moves specialty code and site
A.3) where an episode moves only site (but same specialty)
B) burden on site where transfer originated from
B.1) where an episode arrives to a new site, irrrespective of specialty
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ESAAdmittedAggregated$new()}}
\item \href{#method-specialties}{\code{ESAAdmittedAggregated$specialties()}}
\item \href{#method-transfers}{\code{ESAAdmittedAggregated$transfers()}}
\item \href{#method-longStayPatients_days21}{\code{ESAAdmittedAggregated$longStayPatients_days21()}}
\item \href{#method-longStayPatients_daysAll}{\code{ESAAdmittedAggregated$longStayPatients_daysAll()}}
\item \href{#method-getRawCleanedData}{\code{ESAAdmittedAggregated$getRawCleanedData()}}
\item \href{#method-getProviderSite}{\code{ESAAdmittedAggregated$getProviderSite()}}
\item \href{#method-getDateCol}{\code{ESAAdmittedAggregated$getDateCol()}}
\item \href{#method-clone}{\code{ESAAdmittedAggregated$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ESAAdmittedAggregated$new(
  data,
  date.min,
  date.max,
  episodeStartDate.col,
  episodeEndDate.col,
  dischargeDate.col,
  episodeIdent.col,
  episodeNo.col,
  provider.col,
  providerSite.col,
  patientIdent.col,
  episodeSpecialty.col,
  spellIdent.col = NULL,
  spellHospitalIdent.col = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-specialties"></a>}}
\if{latex}{\out{\hypertarget{method-specialties}{}}}
\subsection{Method \code{specialties()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ESAAdmittedAggregated$specialties()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-transfers"></a>}}
\if{latex}{\out{\hypertarget{method-transfers}{}}}
\subsection{Method \code{transfers()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ESAAdmittedAggregated$transfers()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-longStayPatients_days21"></a>}}
\if{latex}{\out{\hypertarget{method-longStayPatients_days21}{}}}
\subsection{Method \code{longStayPatients_days21()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ESAAdmittedAggregated$longStayPatients_days21()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-longStayPatients_daysAll"></a>}}
\if{latex}{\out{\hypertarget{method-longStayPatients_daysAll}{}}}
\subsection{Method \code{longStayPatients_daysAll()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ESAAdmittedAggregated$longStayPatients_daysAll()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getRawCleanedData"></a>}}
\if{latex}{\out{\hypertarget{method-getRawCleanedData}{}}}
\subsection{Method \code{getRawCleanedData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ESAAdmittedAggregated$getRawCleanedData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getProviderSite"></a>}}
\if{latex}{\out{\hypertarget{method-getProviderSite}{}}}
\subsection{Method \code{getProviderSite()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ESAAdmittedAggregated$getProviderSite()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDateCol"></a>}}
\if{latex}{\out{\hypertarget{method-getDateCol}{}}}
\subsection{Method \code{getDateCol()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ESAAdmittedAggregated$getDateCol()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ESAAdmittedAggregated$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
