% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESAEDAggregated.R
\name{ESAEDAggregated}
\alias{ESAEDAggregated}
\title{Class which acts as a container for ECDS aggregated dataset.
Takes in an ESAEDPatientLevel objects, list of ESASlots, an ESAAcuity Obj
and a queue level
Flags are derived for each of the variables per slot, over the acuity and
queue level}
\description{
Class which acts as a container for ECDS aggregated dataset.
Takes in an ESAEDPatientLevel objects, list of ESASlots, an ESAAcuity Obj
and a queue level
Flags are derived for each of the variables per slot, over the acuity and
queue level

Class which acts as a container for ECDS aggregated dataset.
Takes in an ESAEDPatientLevel objects, list of ESASlots, an ESAAcuity Obj
and a queue level
Flags are derived for each of the variables per slot, over the acuity and
queue level
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ESAEDAggregated$new()}}
\item \href{#method-mergeAdmitted}{\code{ESAEDAggregated$mergeAdmitted()}}
\item \href{#method-getAvailableVars}{\code{ESAEDAggregated$getAvailableVars()}}
\item \href{#method-getIdentifer}{\code{ESAEDAggregated$getIdentifer()}}
\item \href{#method-getAcuity}{\code{ESAEDAggregated$getAcuity()}}
\item \href{#method-getESAPatientLevel}{\code{ESAEDAggregated$getESAPatientLevel()}}
\item \href{#method-getQueue}{\code{ESAEDAggregated$getQueue()}}
\item \href{#method-getSlots}{\code{ESAEDAggregated$getSlots()}}
\item \href{#method-getProviderSite}{\code{ESAEDAggregated$getProviderSite()}}
\item \href{#method-getDate}{\code{ESAEDAggregated$getDate()}}
\item \href{#method-getProvider}{\code{ESAEDAggregated$getProvider()}}
\item \href{#method-clone}{\code{ESAEDAggregated$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ESAEDAggregated$new(
  esaPatientLevel,
  acuity,
  slots,
  provider.col,
  providerSite.col,
  queue = NULL,
  cubicles = NULL,
  cubicles.site = NULL,
  cubicles.date = NULL,
  cubicles.col = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mergeAdmitted"></a>}}
\if{latex}{\out{\hypertarget{method-mergeAdmitted}{}}}
\subsection{Method \code{mergeAdmitted()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ESAEDAggregated$mergeAdmitted(esaAdmittedAggregated)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getAvailableVars"></a>}}
\if{latex}{\out{\hypertarget{method-getAvailableVars}{}}}
\subsection{Method \code{getAvailableVars()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ESAEDAggregated$getAvailableVars()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getIdentifer"></a>}}
\if{latex}{\out{\hypertarget{method-getIdentifer}{}}}
\subsection{Method \code{getIdentifer()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ESAEDAggregated$getIdentifer()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getAcuity"></a>}}
\if{latex}{\out{\hypertarget{method-getAcuity}{}}}
\subsection{Method \code{getAcuity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ESAEDAggregated$getAcuity()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getESAPatientLevel"></a>}}
\if{latex}{\out{\hypertarget{method-getESAPatientLevel}{}}}
\subsection{Method \code{getESAPatientLevel()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ESAEDAggregated$getESAPatientLevel()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getQueue"></a>}}
\if{latex}{\out{\hypertarget{method-getQueue}{}}}
\subsection{Method \code{getQueue()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ESAEDAggregated$getQueue()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getSlots"></a>}}
\if{latex}{\out{\hypertarget{method-getSlots}{}}}
\subsection{Method \code{getSlots()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ESAEDAggregated$getSlots()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getProviderSite"></a>}}
\if{latex}{\out{\hypertarget{method-getProviderSite}{}}}
\subsection{Method \code{getProviderSite()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ESAEDAggregated$getProviderSite()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDate"></a>}}
\if{latex}{\out{\hypertarget{method-getDate}{}}}
\subsection{Method \code{getDate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ESAEDAggregated$getDate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getProvider"></a>}}
\if{latex}{\out{\hypertarget{method-getProvider}{}}}
\subsection{Method \code{getProvider()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ESAEDAggregated$getProvider()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ESAEDAggregated$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
