% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESAEDPatientLevel.R
\name{ESAEDPatientLevel}
\alias{ESAEDPatientLevel}
\title{Title: ESAEDPatientLevel
Author: Edmund Haacke
Date created: 22.09.2021
Date modified: 22.09.2021
Changelog:
- 22.09.21: file created.
Description:
Contain for Emergency Care data patient level}
\description{
Title: ESAEDPatientLevel
Author: Edmund Haacke
Date created: 22.09.2021
Date modified: 22.09.2021
Changelog:
- 22.09.21: file created.
Description:
Contain for Emergency Care data patient level

Title: ESAEDPatientLevel
Author: Edmund Haacke
Date created: 22.09.2021
Date modified: 22.09.2021
Changelog:
- 22.09.21: file created.
Description:
Contain for Emergency Care data patient level
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ESAEDPatientLevel$new()}}
\item \href{#method-getAcuities}{\code{ESAEDPatientLevel$getAcuities()}}
\item \href{#method-getFlags}{\code{ESAEDPatientLevel$getFlags()}}
\item \href{#method-clone}{\code{ESAEDPatientLevel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ESAEDPatientLevel$new(
  data,
  acuities,
  flags,
  slots,
  queues,
  col.arrival,
  col.departure,
  col.uniqueIdent
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getAcuities"></a>}}
\if{latex}{\out{\hypertarget{method-getAcuities}{}}}
\subsection{Method \code{getAcuities()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ESAEDPatientLevel$getAcuities()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getFlags"></a>}}
\if{latex}{\out{\hypertarget{method-getFlags}{}}}
\subsection{Method \code{getFlags()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ESAEDPatientLevel$getFlags()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ESAEDPatientLevel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
