% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESAQueue.R
\name{ESAQueue}
\alias{ESAQueue}
\title{Title: ESAQueue
Author: Edmund Haacke
Date created: 22.09.2021
Date modified: 22.09.2021
Changelog:
- 22.09.21: file created.
Description:}
\description{
Title: ESAQueue
Author: Edmund Haacke
Date created: 22.09.2021
Date modified: 22.09.2021
Changelog:
- 22.09.21: file created.
Description:

Title: ESAQueue
Author: Edmund Haacke
Date created: 22.09.2021
Date modified: 22.09.2021
Changelog:
- 22.09.21: file created.
Description:
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ESAQueue$new()}}
\item \href{#method-clone}{\code{ESAQueue$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ESAQueue$new(name, offset.hours)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ESAQueue$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
