% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESASlot.R
\name{ESASlot}
\alias{ESASlot}
\title{Title: ESASlot
Author: Edmund Haacke
Date created: 22.09.2021
Date modified: 22.09.2021
Changelog:
- 22.09.21: file created.
Description:
Class to define a slot}
\description{
Title: ESASlot
Author: Edmund Haacke
Date created: 22.09.2021
Date modified: 22.09.2021
Changelog:
- 22.09.21: file created.
Description:
Class to define a slot

Title: ESASlot
Author: Edmund Haacke
Date created: 22.09.2021
Date modified: 22.09.2021
Changelog:
- 22.09.21: file created.
Description:
Class to define a slot
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ESASlot$new()}}
\item \href{#method-print}{\code{ESASlot$print()}}
\item \href{#method-clone}{\code{ESASlot$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ESASlot$new(slotID, slotHours)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ESASlot$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ESASlot$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
