% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getDataFromODBC}
\alias{getDataFromODBC}
\title{Load data via an odbc connection}
\usage{
getDataFromODBC(odbcName, odbcQuery)
}
\arguments{
\item{odbcName}{String: The name of the ODBC connection name to utilise}

\item{odbcQuery}{String: The sql query to execute via the ODBC connection}
}
\value{
A data.table of the ODBC connections' response
}
\description{
Load data via an odbc connection
}
