% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.Date}
\alias{is.Date}
\title{Check if an object is date type
is.Date returns TRUE if x is a date}
\usage{
is.Date(x)
}
\arguments{
\item{x}{an R object}
}
\value{
boolean
}
\description{
Check if an object is date type
is.Date returns TRUE if x is a date
}
