% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{searchWithin}
\alias{searchWithin}
\title{Check whether a string/vector is contained within a vector
function which, for a given dataset, checks whether vector, or text, is
contained within another vector.}
\usage{
searchWithin(search, within, quietly = TRUE)
}
\arguments{
\item{search}{A vector or string which is being searched for.}

\item{within}{A vector in which to search for}

\item{quietly}{Boolean whether to print messages}
}
\value{
Boolean: True (Found), False (Not Found)
}
\description{
Check whether a string/vector is contained within a vector
function which, for a given dataset, checks whether vector, or text, is
contained within another vector.
}
