% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unique.vector}
\alias{unique.vector}
\title{Check for comma seperated entries in vector
returns unique vector of individual columns}
\usage{
\method{unique}{vector}(vector)
}
\arguments{
\item{vector}{vector needing checking}
}
\value{

}
\description{
Check for comma seperated entries in vector
returns unique vector of individual columns
}
