% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vector.endsWith}
\alias{vector.endsWith}
\title{Check whether element in x ends in any item in y}
\usage{
vector.endsWith(x, y)
}
\arguments{
\item{x}{vector to search}

\item{y}{vector of endsWiths}
}
\value{
vector of booleans
}
\description{
Check whether element in x ends in any item in y
}
